pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edwh-nostr-messagebus"
dynamic = ["version"]
description = 'EducationWarehouse NOSTR Messagebus'
readme = "README.md"
requires-python = ">=3.11"
license = "MIT"
keywords = ['nostr','education','edwh']
authors = [
  { name = "Remco", email = "remco@educationwarehouse.nl" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
]
dependencies = [
  'monstr',
  'pyinvoke',
  'edwh'
]

[tool.hatch.build.targets.wheel]
packages = ["src/edwh-nostr-messagebus"]
[tool.hatch.build.targets.s-dist]
packages = ["src/edwh-nostr-messagebus"]

[entry-points]
busclient = 'edwh-nostr-messagebus.cli:busclient'

[project.urls]
Documentation = "https://github.com/educationwarehouse/edwh-nostr-messagebus#readme"
Issues = "https://github.com/educationwarehouse/edwh-nostr-messagebus/issues"
Source = "https://github.com/educationwarehouse/edwh-nostr-messagebus"

[tool.hatch.version]
path = "src/edwh-nostr-messagebus/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
  "monstr",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = [
  "black>=23.1.0",
  "mypy>=1.0.0",
  "ruff>=0.0.243",
]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/_ tests}"
style = [
  "ruff {args:.}",
  "black --check --diff {args:.}",
]
fmt = [
  "black {args:.}",
  "ruff --fix {args:.}",
  "style",
]
all = [
  "style",
  "typing",
]

[tool.black]
target-version = ["py311"]
line-length = 120
skip-string-normalization = true

[tool.ruff]
target-version = "py37"
line-length = 120
select = [
  "A",
  "ARG",
  "B",
  "C",
  "DTZ",
  "E",
  "EM",
  "F",
  "FBT",
  "I",
  "ICN",
  "ISC",
  "N",
  "PLC",
  "PLE",
  "PLR",
  "PLW",
  "Q",
  "RUF",
  "S",
  "T",
  "TID",
  "UP",
  "W",
  "YTT",
]
ignore = [
  # Allow non-abstract empty methods in abstract base classes
  "B027",
  # Allow boolean positional values in function calls, like `dict.get(... True)`
  "FBT003",
  # Ignore checks for possible passwords
  "S105", "S106", "S107",
  # Ignore complexity
  "C901", "PLR0911", "PLR0912", "PLR0913", "PLR0915",
]
unfixable = [
  # Don't touch unused imports
  "F401",
]

[tool.ruff.isort]
known-first-party = ["_"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252"]

[tool.coverage.run]
source_pkgs = ["_", "tests"]
branch = true
parallel = true
omit = [
  "src/edwh-nostr-messagebus/__about__.py",
]

[tool.coverage.paths]
_ = ["src/_", "*/-/src/_"]
tests = ["tests", "*/-/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

[tool.hatch.envs.semver]
dependencies = [
    "python-semantic-release",
]
[tool.hatch.envs.semver.scripts]
semver = "semantic-release"
test-semver = [
    'python --version',
    'semantic-release --version',
]
# note: don't run `hatch run publish` right now,
# because semantic release v7 (stable) is not compatible with invoke 2 (which is required for python3.11+)
# and semantic release v8 (alpha) is not compatible with the tool.semantic_release config in this file!
publish = [
    'semantic-release publish',
    'hatch build -c',
    'hatch publish',
]
